bot/requirements.txt
python-telegram-bot==21.0
SQLAlchemy>=2.0.30
PyMySQL==1.1.0
cryptography>=42.0.2
httpx>=0.27
python-dotenv>=1.0.0
pdfplumber>=0.11.0
